Extract and refactor DataGrid into base and derived classes
